try to add arrowheads
For one things, the arrowheads aren't correctly aligned to the line they
are supposed to be drawn on yet.

Also, I cannot figure out how to get the Arrowhead object to actually
show up on the canvas without manually adding a function to get it's
objects and directly return those from an `Objects()` function in a
widget, which defeats the point.